Fix the private yes/no dropdown.
LinkHTMLProperty treats a false value as missing, so if we try to do a
comparison python tries to access __eq__ on the Link object, and because the
base class defines only __cmp__, the Link object __getattr__ is called nand
returns a MissingValue that the __eq__ is look up on, and things go south.

But, we can use 'not', because that calls __nonzero__, which is defined on the
base class.